Change update button style
This commit include two style changes

## Replace button label "update" with "I did it today!"

I want to make what this button do clear.

## Remove `is-small` class

At first, I planned to use this app in a mobile environment. That's
why this `is-small` class makes sense. But, now this app's table
contents is too wide to be used in a mobile environment. So I
remove them.